refactor(editor): table selection runtime 분리
