+++ front/e2e/editor-studio-state.spec.ts
@@ -1253,12 +1253,20 @@ test.describe("editor studio state", () => {
     expect(tableStructureModelSource).toContain("export const canShrinkTableAxisAtEnd = (")
     expect(tableStructureModelSource).toContain("export const countShrinkableTableAxisAtEnd = (")
     expect(tableStructureModelSource).toContain("export const countShrinkableRenderedTableAxisAtEnd = (")
+    expect(tableStructureModelSource).toContain("export const hasTableContextInResolvedPos = (")
+    expect(tableStructureModelSource).toContain("export const isTableSelectionActive = (")
+    expect(tableStructureModelSource).toContain("export const getActiveTableStructureState = (")
     expect(blockEditorEngineSource).toContain('} from "./tableStructureModel"')
     expect(blockEditorEngineSource).not.toContain("const collectSimpleTableColumnCells = (")
     expect(blockEditorEngineSource).not.toContain("const buildReorderedSimpleTableNode = (")
     expect(blockEditorEngineSource).not.toContain("const canShrinkTableAxisAtEnd = (")
     expect(blockEditorEngineSource).not.toContain("const countShrinkableTableAxisAtEnd = (")
     expect(blockEditorEngineSource).not.toContain("const countShrinkableRenderedTableAxisAtEnd = (")
+    expect(blockEditorEngineSource).not.toContain("const TABLE_CONTEXT_NODE_NAMES =")
+    expect(blockEditorEngineSource).not.toContain("const hasTableContextInResolvedPos =")
+    expect(blockEditorEngineSource).not.toContain("const isTableSelectionActive =")
+    expect(blockEditorEngineSource).not.toContain("const getActiveTableStructureState =")
+    expect(blockEditorEngineSource.split("\n").length).toBeLessThan(10600)
     expect(tablePasteModelSource).toContain("export const normalizeTableContextPasteText = (")
     expect(tablePasteModelSource).toContain("const normalizeTableContextPasteLine = (")
     expect(tablePasteModelSource).toContain("const isMarkdownTableRow = (")
@@ -1297,7 +1305,7 @@ test.describe("editor studio state", () => {
     expect(blockEditorEngineSource).toContain("const beginTableAxisDragFromPending = useCallback(")
     expect(blockEditorEngineSource).toContain("const startPendingTableAxisDrag = useCallback(")
     expect(blockEditorEngineSource).toContain("const selectTableAxisAtIndex = useCallback(")
-    expect(blockEditorEngineSource).toContain('overflowMode: getTableOverflowMode(tableNode)')
+    expect(tableStructureModelSource).toContain('overflowMode: getTableOverflowMode(tableNode)')
     expect(tableWidthModelSource).toContain("const maxActiveWidth = Math.max(TABLE_MIN_COLUMN_WIDTH_PX, safeBudget - otherColumnsWidth)")
     expect(blockEditorEngineSource).toContain("const tableCornerGrowSuppressClickRef = useRef(false)")
     expect(blockEditorEngineSource).toContain('"grip" | "grow"')
